pyproject.toml
[project]
name = "shoe-product-discovery"
version = "0.1.0"
description = "Running shoe product discovery with LangChain and Tavily"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "tavily-python>=0.7.19",
    "chainlit>=1.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
